% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_design.R
\name{design_carolina}
\alias{design_carolina}
\title{North carolina design type I and II}
\usage{
design_carolina(set = 2, r = 3, male, female, type = 1)
}
\arguments{
\item{set}{The number of sets}

\item{r}{replication. At least 2 replications is suitable}

\item{male}{females.}

\item{female}{males}

\item{type}{number 1: the North Caroline I. Number 2: the North Caroline II}
}
\description{
North Carolina design I (NCDI) involves the performance of one male which is crossed with more than one females, since the male is selected randomly. It's also known as \code{Polygamous matting design}. On the other hand,
North Carolina design II (NCDII) takes one female at random which is crossed with more than one male. It's also known as \code{Polyandrous matting design}.
}
\examples{
   
res <- design_carolina(set= 3, r =2, male = c("m1","m2"), female = c("f1", "f2"), type = 1)       
}
\references{
Statistical and Biometrical Techniques in Plant Breeding. Jawar R. Sharma. New Age International Publishers 2008.
}
\author{
Omar Benites
}
