% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_design.R
\name{design_lxt}
\alias{design_lxt}
\title{Line by Tester Design
# @param set The number of sets}
\usage{
design_lxt(r = 3, lines, testers, type = 1)
}
\arguments{
\item{r}{replication. At least 2 replications is suitable}

\item{lines}{character vector. Generally, the lines (test genotypes, test != testers) are used as female. In case of employing male stile lines as testers. naturally lines becom the pollen parents.}

\item{testers}{character vector. Generally testers as male or pollen parents. Minimun 2 parents testers and maximum 8-10 parents testers are adequate.}

\item{type}{numeric. 1: parents and full and half siblings. 2: just full and half siblings}
}
\description{
Line by Tester design.
}
\examples{
res <- design_lxt(r =2, lines = c("f1","f2","f3","f4","f5"), testers = c("m1", "m2"), type=1 )
}
\references{
Statistical and Biometrical Techniques in Plant Breeding. Jawar R. Sharma. New Age International Publishers 2008.
}
\author{
Omar Benites
}
